修复调用约定位置写错的问题 将一些类型转换改为显式以减少msvc的警告 将一些用于标识字节数size_t的字段改为uint32_t，反正不会那么大 修复使用msvc时clangd识别不了/std:c++latest编译选项的问题 添加xmake规则以自动更新complie_commands.json 修复Windows上std::filesysytem::path假定字符串使用ansi导致路径转换出错的问题

+++ sources/runtime/src/core/input/input.cpp
@@ -41,10 +41,10 @@ void initalize() {
         return false;
     });
     EventBus::subscribe_event<Display::Events::SysMousePos>(0, [](Display::Events::SysMousePos &e) {
-        mouse_delta_x = e.x - mouse_pos_x;
-        mouse_delta_y = e.y - mouse_pos_y;
-        mouse_pos_x = e.x;
-        mouse_pos_y = e.y;
+        mouse_delta_x = (int32_t)e.x - mouse_pos_x;
+        mouse_delta_y = (int32_t)e.y - mouse_pos_y;
+        mouse_pos_x = (int32_t)e.x;
+        mouse_pos_y = (int32_t)e.y;
         return false;
     });
     EventBus::subscribe_event<Display::Events::SysMouseClick>(0, [](Display::Events::SysMouseClick &e) {
